+++ noodles-vcf/src/writer.rs
@@ -184,7 +184,7 @@ mod tests {
         let mut writer = Writer::new(Vec::new());
 
         let keys: genotypes::Keys = "GT:GQ".parse()?;
-        let genotypes = vec![Genotype::from_str_format("0|0:13", &keys)?];
+        let genotypes = vec![Genotype::from_str_keys("0|0:13", &keys)?];
 
         let record = Record::builder()
             .set_chromosome("sq0".parse()?)